chore: split graphql tests
