+++ test/browser/graphql-api/cookies.mocks.ts
