+++ test/browser/graphql-api/errors.mocks.ts
