+++ test/browser/graphql-api/extensions.mocks.ts
