+++ test/browser/graphql-api/operation-reference.mocks.ts
