+++ test/browser/graphql-api/query.test.ts
@@ -1,6 +1,6 @@
 import { HttpServer } from '@open-draft/test-server/http'
 import { test, expect } from '../playwright.extend'
-import { gql } from '../support/graphql'
+import { gql } from '../../support/graphql'
 
 const EXAMPLE_PATH = require.resolve('./query.mocks.ts')
