+++ test/browser/graphql-api/variables.mocks.ts
