+++ test/graphql-api/response-patching.node.test.ts
@@ -5,10 +5,8 @@ import { graphql } from 'msw'
 import { setupServer } from 'msw/node'
 import fetch from 'cross-fetch'
 import { graphql as executeGraphql, buildSchema } from 'graphql'
-import { ServerApi, createServer } from '@open-draft/test-server'
+import { HttpServer } from '@open-draft/test-server/http'
 import { createGraphQLClient, gql } from '../support/graphql'
-
-let httpServer: ServerApi
 
 const server = setupServer(
   graphql.query('GetUser', async (req, res, ctx) => {
@@ -30,50 +28,52 @@ const server = setupServer(
   }),
 )
 
+const httpServer = new HttpServer((app) => {
+  app.post('/graphql', async (req, res) => {
+    const result = await executeGraphql({
+      schema: buildSchema(gql`
+        type User {
+          firstName: String!
+          lastName: String!
+        }
+
+        # Describing an additional type to return
+        # the request headers back to the request handler.
+        # Apollo will strip off any extra data that
+        # doesn't match the query.
+        type RequestHeader {
+          name: String!
+          value: String!
+        }
+
+        type Query {
+          user: User!
+          requestHeaders: [RequestHeader!]
+        }
+      `),
+      operationName: 'GetUser',
+      source: req.body.query,
+      rootValue: {
+        user: {
+          firstName: 'John',
+          lastName: 'Maverick',
+        },
+      },
+    })
+
+    return res.status(200).json({
+      requestHeaders: req.headers,
+      queryResult: result,
+    })
+  })
+})
+
 beforeAll(async () => {
   server.listen()
 
   // This test server acts as a production server MSW will be hitting
   // when performing a request patching with `ctx.fetch()`.
-  httpServer = await createServer((app) => {
-    app.post('/graphql', async (req, res) => {
-      const result = await executeGraphql({
-        schema: buildSchema(gql`
-          type User {
-            firstName: String!
-            lastName: String!
-          }
-
-          # Describing an additional type to return
-          # the request headers back to the request handler.
-          # Apollo will strip off any extra data that
-          # doesn't match the query.
-          type RequestHeader {
-            name: String!
-            value: String!
-          }
-
-          type Query {
-            user: User!
-            requestHeaders: [RequestHeader!]
-          }
-        `),
-        operationName: 'GetUser',
-        source: req.body.query,
-        rootValue: {
-          user: {
-            firstName: 'John',
-            lastName: 'Maverick',
-          },
-        },
-      })
-
-      return res.status(200).json({
-        requestHeaders: req.headers,
-        queryResult: result,
-      })
-    })
-  })
+  await httpServer.listen()
 })
 
 afterAll(async () => {
@@ -83,7 +83,7 @@ afterAll(async () => {
 
 test('patches a GraphQL response', async () => {
   const client = createGraphQLClient({
-    uri: httpServer.http.makeUrl('/graphql'),
+    uri: httpServer.http.url('/graphql'),
     fetch,
   })
